handle model and cache name prefixes
